Give a pending composite presentation a single owner
`usePresentItem` moves to the composite utils and becomes the one owner
of the presentations a component asks for, so no call site tracks them
itself. It records the store it is presenting for in a layout effect and
refuses any request whose store is not the current owner.

The layout cleanup runs while the unmount is still being committed,
before the microtask the focus handler queued, so a request asked for
once the component is gone is refused rather than created. The passive
cleanup it replaces ran at that point too, but it had nothing to cancel
yet, and a passive setup could not have restored ownership either, since
the composite's own setup runs after the child that asks for the
effect-driven presentation.

Recording the store rather than a mounted flag keeps cancellation scoped
across a `store` prop swap. `CompositeItem` uses the same hook now, in
place of bookkeeping that recorded no store identity and had no cleanup,
and `presentItem` is no longer exported.

Fixes https://github.com/ariakit/ariakit/issues/7024
