+++ .changeset/300-composite-presentation-owner.md
@@ -1,0 +1,6 @@
+---
+"@ariakit/react-components": patch
+"@ariakit/react": patch
+---
+
+Fixed [`Composite`](https://ariakit.com/reference/composite) moving DOM focus on its own after being unmounted while it was focused.
